+++ primitives/bilinear.hpp
@@ -1,6 +1,8 @@
 #ifndef BILINEAR_HPP
 #define BILINEAR_HPP
 
+#include "numtype.h"
+
 #include <vector>
 #include "vector.h"
 #include "grid.hpp"
@@ -25,16 +27,16 @@ class Bilinear: public Surface {
         BBox bbox;
         bool has_bounds;
         
-        float last_rayw;
+        float32 last_rayw;
     
-        Bilinear(unsigned int res_time_);
+        Bilinear(uint8 res_time_);
         Bilinear(Vec3 v1, Vec3 v2, Vec3 v3, Vec3 v4);
         virtual ~Bilinear();
         
         virtual bool intersect_ray(Ray &ray, Intersection *intersection=NULL);
         virtual BBox &bounds();
-        virtual int dice_rate(float upoly_width);
-        bool is_traceable(float ray_width);
+        virtual int dice_rate(float32 upoly_width);
+        bool is_traceable(float32 ray_width);
         virtual void refine(std::vector<Primitive *> &primitives);
         
         void add_time_sample(int samp, Vec3 v1, Vec3 v2, Vec3 v3, Vec3 v4);
